split out scheduler
