+++ broccoli_server/application.py
@@ -9,14 +9,11 @@ from broccoli_server.database import Migration
 from broccoli_server.utils import validate_schema_or_not, getenv_or_raise
 from broccoli_server.utils.request_schemas import ADD_WORKER_BODY_SCHEMA
 from broccoli_server.content import ContentStore
-from broccoli_server.scheduler import WorkerConfigStore
-from broccoli_server.scheduler import GlobalMetadataStore
-from broccoli_server.scheduler import Reconciler
-from broccoli_server.scheduler import Worker
+from broccoli_server.worker import WorkerConfigStore, GlobalMetadataStore, Worker, WorkerCache
+from broccoli_server.reconciler import Reconciler
 from broccoli_server.mod_view import ModViewStore
 from broccoli_server.mod_view import ModViewRenderer
 from broccoli_server.mod_view import ModViewQuery
-from broccoli_server.scheduler import WorkerCache
 from flask import Flask, request, jsonify, send_from_directory, redirect
 from flask_cors import CORS
 from flask_jwt_extended import JWTManager, create_access_token, verify_jwt_in_request
