+++ broccoli_server/reconciler/logging.py
@@ -1,0 +1,6 @@
+import logging
+from broccoli_server.utils import DefaultHandler, get_logging_level
+
+logger = logging.getLogger('reconciler')
+logger.setLevel(get_logging_level())
+logger.addHandler(DefaultHandler)
